refactor: extract systemd checks into own module
The systemd utils module still exists because the XOrgIdleTime check
also optionally uses systemd.
