+++ src/autosuspend/checks/wakeup.py
@@ -1,15 +1,13 @@
-import configparser
+import configparser  # noqa
 from contextlib import suppress
 from datetime import datetime, timedelta, timezone
 from pathlib import Path
-import re
 import subprocess
-from typing import Optional, Pattern
+from typing import Optional
 
 from . import ConfigurationError, TemporaryCheckError, Wakeup
 from .util import CommandMixin
 from ..util.subprocess import raise_severe_if_command_not_found
-from ..util.systemd import next_timer_executions
 
 
 # isort: off
@@ -19,6 +17,8 @@ with suppress(ModuleNotFoundError):
 with suppress(ModuleNotFoundError):
     from .xpath import XPathWakeup as XPath  # noqa
     from .xpath import XPathDeltaWakeup as XPathDelta  # noqa
+with suppress(ModuleNotFoundError):
+    from .systemd import SystemdTimer  # noqa
 
 # isort: on
 
@@ -110,28 +110,3 @@ class Periodic(Wakeup):
 
     def check(self, timestamp: datetime) -> Optional[datetime]:
         return timestamp + self._delta
-
-
-class SystemdTimer(Wakeup):
-    """Ensures that the system is active when some selected SystemD timers will run."""
-
-    @classmethod
-    def create(cls, name: str, config: configparser.SectionProxy) -> "SystemdTimer":
-        try:
-            return cls(name, re.compile(config["match"]))
-        except (re.error, ValueError, KeyError, TypeError) as error:
-            raise ConfigurationError(str(error))
-
-    def __init__(self, name: str, match: Pattern) -> None:
-        Wakeup.__init__(self, name)
-        self._match = match
-
-    def check(self, timestamp: datetime) -> Optional[datetime]:
-        executions = next_timer_executions()
-        matching_executions = [
-            next_run for name, next_run in executions.items() if self._match.match(name)
-        ]
-        try:
-            return min(matching_executions)
-        except ValueError:
-            return None
