+++ tests/test_checks_wakeup.py
@@ -1,8 +1,6 @@
 from datetime import datetime, timedelta, timezone
 from pathlib import Path
-import re
 import subprocess
-from unittest.mock import Mock
 
 import pytest
 from pytest_mock import MockFixture
@@ -13,7 +11,7 @@ from autosuspend.checks import (
     SevereCheckError,
     TemporaryCheckError,
 )
-from autosuspend.checks.wakeup import Command, File, Periodic, SystemdTimer
+from autosuspend.checks.wakeup import Command, File, Periodic
 
 from . import CheckTest
 from .utils import config_section
@@ -143,51 +141,3 @@ class TestPeriodic(CheckTest):
         check = Periodic("test", delta)
         now = datetime.now(timezone.utc)
         assert check.check(now) == now + delta
-
-
-class TestSystemdTimer(CheckTest):
-    @staticmethod
-    @pytest.fixture()
-    def next_timer_executions(mocker: MockFixture) -> Mock:
-        return mocker.patch("autosuspend.checks.wakeup.next_timer_executions")
-
-    def create_instance(self, name: str) -> Check:
-        return SystemdTimer(name, re.compile(".*"))
-
-    def test_create_handles_incorrect_expressions(self) -> None:
-        with pytest.raises(ConfigurationError):
-            SystemdTimer.create("somename", config_section({"match": "(.*"}))
-
-    def test_create_raises_if_match_is_missing(self) -> None:
-        with pytest.raises(ConfigurationError):
-            SystemdTimer.create("somename", config_section())
-
-    def test_works_without_timers(self, next_timer_executions: Mock) -> None:
-        next_timer_executions.return_value = {}
-        now = datetime.now(timezone.utc)
-
-        assert SystemdTimer("foo", re.compile(".*")).check(now) is None
-
-    def test_ignores_non_matching_timers(self, next_timer_executions: Mock) -> None:
-        now = datetime.now(timezone.utc)
-        next_timer_executions.return_value = {"ignored": now}
-
-        assert SystemdTimer("foo", re.compile("needle")).check(now) is None
-
-    def test_finds_matching_timers(self, next_timer_executions: Mock) -> None:
-        pattern = "foo"
-        now = datetime.now(timezone.utc)
-        next_timer_executions.return_value = {pattern: now}
-
-        assert SystemdTimer("foo", re.compile(pattern)).check(now) is now
-
-    def test_selects_the_closest_execution_if_multiple_match(
-        self, next_timer_executions: Mock
-    ) -> None:
-        now = datetime.now(timezone.utc)
-        next_timer_executions.return_value = {
-            "later": now + timedelta(minutes=1),
-            "matching": now,
-        }
-
-        assert SystemdTimer("foo", re.compile(".*")).check(now) is now
