+++ tests/test_util_systemd.py
@@ -1,11 +1,7 @@
 from dbus.proxies import ProxyObject
 import pytest
 
-from autosuspend.util.systemd import (
-    list_logind_sessions,
-    LogindDBusException,
-    next_timer_executions,
-)
+from autosuspend.util.systemd import list_logind_sessions, LogindDBusException
 
 
 def test_list_logind_sessions_empty(logind: ProxyObject) -> None:
@@ -21,8 +17,3 @@ def test_list_logind_sessions_empty(logind: ProxyObject) -> None:
 def test_list_logind_sessions_dbus_error() -> None:
     with pytest.raises(LogindDBusException):
         list_logind_sessions()
-
-
-@pytest.mark.skip(reason="No dbusmock implementation available")
-def test_next_timer_executions() -> None:
-    assert next_timer_executions() is not None
